+++ src/dataStructures/tree/avlTree/types/index.ts
@@ -2,6 +2,5 @@ import { BinarySearchTreeInterface,
   BinarySearchTreeNodeInterface } from '../../binarySearchTree/types';
 
 export interface AvlTreeInterface<T = unknown> extends BinarySearchTreeInterface<T> {
-  readonly binarySearchTree: BinarySearchTreeInterface<T>;
   balance: (node: BinarySearchTreeNodeInterface<T>) => this;
 }