+++ src/dataStructures/tree/redBlackRree/types/index.ts
@@ -18,7 +18,6 @@ export interface RedBlackTreeNodeInterface<T = unknown> {
 export interface RedBlackTreeInterface<T = unknown> {
   readonly comparator: Comparator;
   readonly root: CompleteRedBlackTreeNode<T> | null;
-  readonly binarySearchTree: BinarySearchTreeInterface<RedBlackTreeNodeInterface<T>>;
   setRoot: (root: CompleteRedBlackTreeNode<T> | null) => this;
   find: (value: T) => null | CompleteRedBlackTreeNode<T>;
   findMin: () => null | CompleteRedBlackTreeNode<T>;